+++ packages/calcite-components/src/components/table/table.e2e.ts
@@ -11,7 +11,7 @@ describe("calcite-table", () => {
   describe("renders", () => {
     renders(
       html`<calcite-table caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -94,7 +94,7 @@ describe("calcite-table", () => {
     describe("is accessible simple", () => {
       accessible(
         html`<calcite-table caption="Simple table">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -117,7 +117,7 @@ describe("calcite-table", () => {
     describe("is accessible with selection mode multiple", () => {
       accessible(
         html`<calcite-table caption="Simple table" selection-mode="multiple">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -140,7 +140,7 @@ describe("calcite-table", () => {
     describe("is accessible with selection mode multiple selected at load", () => {
       accessible(
         html`<calcite-table caption="Simple table" selection-mode="multiple">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -163,7 +163,7 @@ describe("calcite-table", () => {
     describe("is accessible with selection mode single", () => {
       accessible(
         html`<calcite-table caption="Simple table" selection-mode="single">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -186,7 +186,7 @@ describe("calcite-table", () => {
     describe("is accessible with numbered", () => {
       accessible(
         html`<calcite-table caption="Simple table" numbered>
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -209,7 +209,7 @@ describe("calcite-table", () => {
     describe("is accessible with numbered and selection", () => {
       accessible(
         html`<calcite-table caption="Simple table" numbered selection-mode="multiple">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -232,7 +232,7 @@ describe("calcite-table", () => {
     describe("is accessible with pagination", () => {
       accessible(
         html`<calcite-table page-size="4" caption="Simple table">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -271,7 +271,7 @@ describe("calcite-table", () => {
     describe("is accessible with pagination and interaction mode static", () => {
       accessible(
         html`<calcite-table page-size="4" caption="Simple table" interaction-mode="static">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -310,7 +310,7 @@ describe("calcite-table", () => {
     describe("is accessible with pagination and selection mode", () => {
       accessible(
         html`<calcite-table page-size="4" selection-mode="multiple" caption="Simple table">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
             <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           </calcite-table-row>
@@ -353,7 +353,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="single" caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -468,7 +468,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -583,7 +583,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="single" caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -693,7 +693,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -802,7 +802,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row slot=${SLOTS.tableHeader}>
+        <calcite-table-row slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -854,7 +854,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -906,7 +906,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table" page-size="1">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -958,7 +958,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1013,7 +1013,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1068,7 +1068,7 @@ describe("selection modes", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table" page-size="2" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1134,7 +1134,7 @@ describe("pagination event", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table" page-size="1" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1208,7 +1208,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Simple table" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1257,20 +1257,24 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2a");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3b");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
   });
@@ -1279,7 +1283,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Simple table" page-size="2" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1336,12 +1340,14 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2b");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2b");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
   });
@@ -1350,7 +1356,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Simple table" page-size="2" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1407,12 +1413,14 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2b");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2b");
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
 
@@ -1458,12 +1466,14 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-4b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-4b");
 
@@ -1510,12 +1520,14 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-5b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-5b");
   });
@@ -1524,7 +1536,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1577,7 +1589,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1627,8 +1639,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("PageUp");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3b");
   });
@@ -1637,11 +1650,11 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Multiple headers using col-span table">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header id="head-1a" col-span="2" heading="Name"></calcite-table-header>
             <calcite-table-header id="head-1b" col-span="2" heading="Information"></calcite-table-header>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header id="head-2a" heading="First"></calcite-table-header>
             <calcite-table-header id="head-2b" heading="Last"></calcite-table-header>
             <calcite-table-header id="head-2c" heading="Education level"></calcite-table-header>
@@ -1663,13 +1676,13 @@ describe("keyboard navigation", () => {
             <calcite-table-cell id="cell-4a">cell</calcite-table-cell>
             <calcite-table-cell id="cell-4b" col-span="3">cell</calcite-table-cell>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableFooter}>
+          <calcite-table-row slot="${SLOTS.tableFooter}">
             <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1c">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1d">foot</calcite-table-cell>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableFooter}>
+          <calcite-table-row slot="${SLOTS.tableFooter}">
             <calcite-table-cell id="foot-2a" col-span="2">foot</calcite-table-cell>
             <calcite-table-cell id="foot-2b" col-span="2">foot</calcite-table-cell>
           </calcite-table-row>
@@ -1710,8 +1723,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-2d");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-2b");
     await page.keyboard.press("PageUp");
@@ -1726,8 +1740,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("ArrowDown");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
   });
@@ -1736,11 +1751,11 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table caption="Multiple headers using col-span table" page-size="2" style="width:800px">
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header id="head-1a" col-span="2" heading="Name"></calcite-table-header>
             <calcite-table-header id="head-1b" col-span="2" heading="Information"></calcite-table-header>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableHeader}>
+          <calcite-table-row slot="${SLOTS.tableHeader}">
             <calcite-table-header id="head-2a" heading="First"></calcite-table-header>
             <calcite-table-header id="head-2b" heading="Last"></calcite-table-header>
             <calcite-table-header id="head-2c" heading="Education level"></calcite-table-header>
@@ -1762,13 +1777,13 @@ describe("keyboard navigation", () => {
             <calcite-table-cell id="cell-4a">cell</calcite-table-cell>
             <calcite-table-cell id="cell-4b" col-span="3">cell</calcite-table-cell>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableFooter}>
+          <calcite-table-row slot="${SLOTS.tableFooter}">
             <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1c">foot</calcite-table-cell>
             <calcite-table-cell id="foot-1d">foot</calcite-table-cell>
           </calcite-table-row>
-          <calcite-table-row slot=${SLOTS.tableFooter}>
+          <calcite-table-row slot="${SLOTS.tableFooter}">
             <calcite-table-cell id="foot-2a" col-span="2">foot</calcite-table-cell>
             <calcite-table-cell id="foot-2b" col-span="2">foot</calcite-table-cell>
           </calcite-table-row>
@@ -1815,8 +1830,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-2d");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-2b");
     await page.keyboard.press("Home");
@@ -1837,12 +1853,14 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-2a");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-2b");
 
@@ -1892,8 +1910,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-2d");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-2b");
     await page.keyboard.press("Home");
@@ -1917,8 +1936,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("End");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-2b");
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("head-1a");
     await page.keyboard.press("ArrowDown");
@@ -1933,7 +1953,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table selection-mode="multiple" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -1949,7 +1969,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-3a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-3b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>
@@ -1964,10 +1984,9 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
 
     expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot?.activeElement.shadowRoot?.querySelector("th").classList,
-      ),
+      await page.$eval(`#${rowHead.id}`, async (el) => {
+        return el.shadowRoot?.activeElement.shadowRoot?.querySelector("th").classList;
+      }),
     ).toEqual({ "0": CSS.selectionCell, "1": CSS.multipleSelectionCell });
 
     await page.keyboard.press("ArrowRight");
@@ -2006,13 +2025,15 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3a");
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-1b");
 
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(
       await page.$eval(
@@ -2026,7 +2047,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table numbered caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -2042,7 +2063,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-3a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-3b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>
@@ -2099,13 +2120,15 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3a");
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-1b");
 
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(
       await page.$eval(
@@ -2119,7 +2142,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table numbered selection-mode="single" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -2135,7 +2158,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-3a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-3b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>
@@ -2206,8 +2229,9 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-3a");
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(
       await page.$eval(
@@ -2221,7 +2245,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table numbered selection-mode="multiple" page-size="2" caption="Simple table">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -2241,7 +2265,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-4a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-4b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>
@@ -2304,8 +2328,9 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("cell-2a");
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-1b");
 
@@ -2318,8 +2343,9 @@ describe("keyboard navigation", () => {
       ),
     ).toEqual({ "0": CELL_CSS.footerCell, "1": CSS.numberCell });
 
-    page.keyboard.press("ControlLeft");
+    await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
     expect(
       await page.$eval(
@@ -2333,7 +2359,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table numbered selection-mode="single" page-size="2" caption="Simple table" style="width:800px">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -2353,7 +2379,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-4a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-4b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>
@@ -2371,10 +2397,9 @@ describe("keyboard navigation", () => {
     await page.waitForChanges();
 
     expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot?.activeElement.shadowRoot?.querySelector("th").classList,
-      ),
+      await page.$eval(`#${rowHead.id}`, (el) => {
+        return el.shadowRoot?.activeElement.shadowRoot?.querySelector("th").classList;
+      }),
     ).toEqual({ "0": CSS.numberCell });
 
     await page.keyboard.press("ArrowRight");
@@ -2433,8 +2458,9 @@ describe("keyboard navigation", () => {
       await page.$eval(`#${row2.id}`, (el) => el.shadowRoot?.activeElement.shadowRoot?.querySelector("td").classList),
     ).toEqual({ "0": CSS.selectionCell });
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-1b");
 
@@ -2447,9 +2473,9 @@ describe("keyboard navigation", () => {
       ),
     ).toEqual({ "0": CELL_CSS.footerCell, "1": CSS.numberCell });
 
-    page.keyboard.press("ControlRight");
-
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(
       await page.$eval(
@@ -2519,8 +2545,9 @@ describe("keyboard navigation", () => {
       await page.$eval(`#${row4.id}`, (el) => el.shadowRoot?.activeElement.shadowRoot?.querySelector("td").classList),
     ).toEqual({ "0": CSS.numberCell });
 
-    page.keyboard.press("ControlRight");
+    await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
+    await page.keyboard.up("ControlRight");
     await page.waitForChanges();
     expect(await getFocusedElementProp(page, "id")).toBe("foot-1b");
 
@@ -2538,7 +2565,7 @@ describe("keyboard navigation", () => {
     const page = await newE2EPage();
     await page.setContent(
       html`<calcite-table numbered selection-mode="multiple" caption="Simple table" interaction-mode="static">
-        <calcite-table-row id="row-head" slot=${SLOTS.tableHeader}>
+        <calcite-table-row id="row-head" slot="${SLOTS.tableHeader}">
           <calcite-table-header id="head-1a" heading="Heading" description="Description"></calcite-table-header>
           <calcite-table-header id="head-1b" heading="Heading" description="Description"></calcite-table-header>
         </calcite-table-row>
@@ -2554,7 +2581,7 @@ describe("keyboard navigation", () => {
           <calcite-table-cell id="cell-3a">cell</calcite-table-cell>
           <calcite-table-cell id="cell-3b">cell</calcite-table-cell>
         </calcite-table-row>
-        <calcite-table-row slot=${SLOTS.tableFooter} id="row-foot">
+        <calcite-table-row slot="${SLOTS.tableFooter}" id="row-foot">
           <calcite-table-cell id="foot-1a">foot</calcite-table-cell>
           <calcite-table-cell id="foot-1b">foot</calcite-table-cell>
         </calcite-table-row>