+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.0-next.7](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.6...@esri/calcite-components-react@3.0.0-next.7) (2024-11-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.6](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.5...@esri/calcite-components-react@3.0.0-next.6) (2024-11-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.0.0-next.5](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.4...@esri/calcite-components-react@3.0.0-next.5) (2024-11-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.0.0-next.4](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.0.0-next.3...@esri/calcite-components-react@3.0.0-next.4) (2024-10-31)
 
 **Note:** Version bump only for package @esri/calcite-components-react